requirements.txt
discord.py
yfinance
psutil
python-dotenv
requests
beautifulsoup4
